requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.1.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
Authlib==1.3.0
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
DateTime==5.4
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
distro==1.8.0
dm-tree==0.1.8
docstring-parser==0.15
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
google-pasta==0.2.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonify==0.5
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.0.1
kfp-server-api==2.0.5
kiwisolver==1.4.5
kubernetes==28.1.0
libclang==16.0.6
linkify-it-py==2.0.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.3.1
mpmath==1.3.0
msal==1.26.0
namex==0.0.7
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
openai==1.4.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
panel==1.3.4
param==2.0.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.16.0
xyzservices==2023.10.1
zope.interface==6.1
